+++ packages/core/test/style_text_align_test.dart
@@ -394,7 +394,7 @@ void main() {
       equals(
         '[SizedBox:0.0x5.0],'
         '[CssBlock:child='
-        '[Padding:(0,5,0,5),child='
+        '[HorizontalMargin:left=5,right=5,child='
         '[CssBlock:child=[RichText:align=center,(:Foo)]]'
         ']],'
         '[SizedBox:0.0x5.0]',